+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 # Changelog
 
+- Add sort by filter in explore section (roubkar)
 - Close select form suggestions on ESC key press (roubkar)
 - Fix run navigation link issue (roubkar)
 - Rerender parallel coordinates plot on chart settings update (roubkar)